+++ tidy3d/plugins/smatrix/component_modelers/base.py
@@ -22,7 +22,8 @@ from tidy3d.plugins.smatrix.ports.coaxial_lumped import CoaxialLumpedPort
 from tidy3d.plugins.smatrix.ports.modal import Port
 from tidy3d.plugins.smatrix.ports.rectangular_lumped import LumpedPort
 from tidy3d.plugins.smatrix.ports.wave import WavePort
-from tidy3d.web.api.container import Batch, BatchData
+from tidy3d.web.api.batch_data import BatchData
+from tidy3d.web.api.container import Batch
 
 # fwidth of gaussian pulse in units of central frequency
 FWIDTH_FRAC = 1.0 / 10
